requirements.txt
numpy~=1.25.0
pandas~=2.0.3
Pillow~=9.5.0
Flask~=2.3.2
Flask-Bootstrap==3.3.7.1
Jinja2==3.1.2
gunicorn==20.1.0
pip==23.1.2